P824-002 demangling of generic subprograms
When rewriting of instances of generic subprograms we were demangling name
of the first subprogram occurring in the wrapper package. This was wrong for
instances of generic subprograms whose actuals are another subprograms, as
they occur first in the wrapper package. Now we demangle name of the first
subprogram in the wrapper package that is not a generic actual subprogram.

This only seem to affect mapping between entity names and entity ids, which
in turn only affect analysis of subprograms not in SPARK; no test affected.
Perhaps it also affects the "R" suffix in flow messages on generic
subprograms instantiated with other subprograms, but this is a small issue.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Subprogram_Instantiation): fixed for generic subprograms
 instantiated with actual subprograms.

Change-Id: Icf4f07320565b7484e0f19a31886aee195d9e30b