P801-002 refactor build of predicates on composite types
We introduce generic predicates which can be used to
lift a predicate on a component of a composite type to
the composite type.

* spark_util-types.ads
(Get_Stored_Constraint_For_Discr): Search for the
constraint stored for a given discriminant in a
constrained record type.

* gnat2why-expr.adb
(Compute_Default_Init): Refactor to use functions that
build predicates for composite types.
(Compute_Dynamic_Invariant): Idem.

* why-gen-arrays.ads
(Build_Predicate_For_Array): Generic function that builds
predicates for components of an array.
* why-gen-records.ads
(Build_Predicate_For_Record): Generic function that builds
predicates for components of a record.

Change-Id: Id18ac4533f183c910950b642ef6553b193d3573d