P825-006 Missing range check on inlined code
In some cases, the special inlining performed in GNATprove for local
functions without contracts may lead to a range check being missed on
the call to such a function. This is because a compiler-generated
unchecked conversion is created with the range check flag set, a case
which was badly handled by GNATprove. Now fixed.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Unchecked_Type_Conversion): Rewrite some compiler-generated
  unchecked type conversions. This ensures that remaining
  compiler-generated type conversions can be skipped when translating
  into Why3. We cannot simply rewrite all of them, as the presence of
  the type conversion is required e.g. for translating literals.

* gnat2why/why/gnat2why-expr.adb
(Check_Scalar_Range): Frontend does not always issue an error on static
  values converted outside of their range, in particular not on
  compiler-generated conversions. So we cannot rely on the value being
  static to skip a check. Instead, we check that the value is in range to
  decide to skip the check.

* testsuite/gnatprove/tests/P825-006__inlined_range/main.adb
* testsuite/gnatprove/tests/P825-006__inlined_range/test.out
* testsuite/gnatprove/tests/P825-006__inlined_range/test.py
Add test.

Change-Id: Ia7cca5113bd66b759d0060d2768bc922e7b5756e