P906-011 fix missing division check on float exponentiation
When the exponent is negative, a float exponentiation is equal to
reciprocal of the same exponentiation with a positive exponent. But this
division may fail if the base is zero. We add a check for this situation
here. Currently, we do not try to detect statically if the exponent is
negative or not.

* gnat2why-expr.adb
(New_Op_Expr): add a division check for float exponentiation, which
checks that the base is not 0.0.

Change-Id: I59e2e20241178bcaaac83e48cae394a4cd010847