P801-002 introduce tool limitations for invariants in marking
We won't attempt to support type invariants on tagged or
protected types, nor on types declared in a nested package.

* gnat2why/spark/spark_definition.adb
(Mark_Entity): Mark type invariants on tagged types, protected
types and types declared in nested packages as unsupported.

* testsuite/gnatprove/tests/P829-014__invariant/test.out
The test cannot be rewritten without an invariant in a nested
package.

Change-Id: I5d5ec3dcfc23dd8bac2a3dc1cae6e56117b1d287