P910-007 do not register tasks in nested pkgs
Task objects declared in nested packages were wrongly registered twice:
first time when processing elaboration of the nested package and second time
when processing elaboration of the enclosing package. Now we only register
tasks declared directly in the scope of the analyzed entity (which must be
a package, because that's the only place where tasks and protected objects
may be declared in Ravenscar).

This also fixes repeated registering of priorities of protected objects,
which was not very useful but did not make any harm.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Object_Declaration): register library-level tasks and protected objects
  only when declared directly in the scope of the analyzed entity.

* testsuite/gnatprove/tests/P910-007__task_nested_pkg/test.out
Output for primary test updated.

* testsuite/gnatprove/tests/P820-002__protected_out/test.out
Output updated (spurious errors removed), xref P820-002.

Change-Id: Idaa66a625e2af873b3d82c3aae3b7165d3b7f3ad