P907-023 flow - improve treatment of quantified variables
Get_Variables now does not return quantified variables in most cases:

  Get_Variables ( "for all I in 1 .. X => A (I)" )  will now return {A, X}

But note that if you call it from a context that is not aware of
the quantifier then you still get them (this is by design), i.e:

  Get_Variables ( "A (I)" )                         will return {A, I}

There is a massive amount of code in Flow-CFG that can now be removed
that dealt with this problem: we used to have a separate tree traversal
as a pre-processing step that picked up all the quantified variables, but
this has all been deleted now.

I have also taken the opportunity to clean up Get_Variables. There are
more improvements possible (in particular treatment of attributes and
bounds), but this is a good first step.

We now also have a better postcondition on Get_Globals to clarify that
the variables returned are always entire variables.

Change-Id: Ia06fc2a03b8246209b76acd12c3f3b0333ef7c44