minor cleanup of Vertex_Vectors
Type Vertex_Vectors.Vector was never used with a random access, so it could
be a list. Vector-specific operation was only used to concatenate before
iterating, which hardly justifies the use of vector container.

This patch changes Vector_Vectors to list. The only non-trivial change is
the already mentioned iteration, where we (destructively) join two lists
with Splice and then iterate over them with a single FOR loop. This avoid
creating copies of the contents of the initial lists.

Also, now the cluster id for the first vertex corresponding to the called
suprogram is set only once, not twice.

* gnat2why/flow/flow-control_flow_graph.adb
(Vertex_Lists): changed from Vertex_Vector.

(Do_Call_Statement): rewritten with splicing.

Change-Id: Ie22b343ae6a728b181da3739dcf44250d38f581e