P801-002 support type invariants in proof
* common/vc_kinds.ads
Add two VC kinds for invariant checks.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Tagged types with a component that has a type
invariant are not allowed in the immediate scope of the type.

* gnat2why/spark/spark_util-subprograms.ads
(Subp_Needs_Invariant_Checks): Return True whenever an invariant
check is needing when calling a subprogram.

* gnat2why/spark/spark_util-types.ads
(Has_Invariants_In_SPARK): Returns True for types with a type
invariant if the invariant is in SPARK.
(Has_Visible_Type_Invariants): Returns True for types with a top
level invariant which needs to be checked in the current
compilation unit.
(Invariant_Check_Needed): Returns True if there is an invariant
that needs to be checked for the type. It can come from the type
itself, from one of its ancestor, or from one of its components.

* gnat2why/why/gnat2why-error_messages.adb
(Not_Proved_Message): Add messages for invariant checks.
(Proved_Message): Add messages for invariant checks.

* gnat2why/why/gnat2why-expr.adb
(New_Type_Invariant_Call): Constructs a call to the why3 predicate
for a top-level invariant.
(New_Invariant_Check): Returns a program that performs a type
invariant check on an expression.
(Type_Invariant_Expression): Returns the translation of the
expression contained in an invariant procedure.
(Compute_Default_Init): New argument Include_Subtypes which
defaults to True and allows to assume default initailization of a
variable of an unconstrained type.
(Compute_Dynamic_Invariant): Assume the type invariant if we are
not in the immediate scope of the type.
(Compute_Top_Level_Type_Invariant): Returns a predicate for the
top level invariant of a type if any.
(Compute_Type_Invariant): Compute a predicate including all the
type invariants that apply to an expression. A parameter allows
to include invariants declared in or outside the current
compilation unit.
(Insert_Invariant_Check): Reconstruct the binder of an expression
before performing an invariant check.
(Transform_Declaration): Check type invariants on declaration of
visible constants without variable inputs. On a type declaration,
check that there can be no runtime error in the invariant and that
the invariant holds after default initialization.
(Transform_Expr): Do an invariant check on conversions.
(Transform_Function_Call): Do an invariant check before the call.
(Transform_Statement_Or_Declaration): Do an invariant check
before procedure calls.
(Variables_In_Dynamic_Invariant): Take into account the variables
used in type invariants declared outside the current compilation
unit.
(Variables_In_Type_Invariant): Return the set of variables occuring
in a top level type invariant.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Type_Invariants_For_Subprogram): Computes a predicate
which performs every invariant check required when calling or
returning from a subprogram.
(Compute_Type_Invariants_For_Package): Computes a predicate
which performs every invariant check required after the elaboration
of a package.
(Generate_VCs_For_Package_Elaboration): Check the type invariants
of variables and constants with variables inputs that are mentioned
in the package's Initializes aspect.
(Generate_VCs_For_Subprogram): Assume the type invariants of global
inputs and of in parameters if the subprogram is boundary and check
the type invariants of global outputs and of out parameters if the
subprogram is boundary.
(Generate_Axiom_For_Post): Add the type invariants on inputs to the
guards of the axiom and assume the type invariants of the result.
(Generate_Subprogram_Program_Fun): Add the type invariants on
outputs to the subprogram postcondition and declare an additional
program function which should be used to check the type invariants
on inputs.

* gnat2why/why/gnat2why-types.adb
(Generate_Type_Completion): Generates a predicate function for the
top level type invariant of the type if any.

* gnat2why/why/why-atree-modules.adb
(Insert_Why_Symbols): Insert the symbol for the program function
used to check invariants on subprogram calls.

* gnat2why/why/why-gen-names.ads
New names for the invariant predicate and the invariant checking
program function.

* why3
Support new checks for type invariants.

Change-Id: I5d911f90e3e45053a5441acc46cfb61c80b7a5d4