P525-016 detect of null default subprogram
When detecting formals of subprograms coming from expansion of null default
parameters of generic units we relied on the fact that their body contains
only a null statement (followed by a single return statement for null
functions). This was checked by Has_Null_Body.

I think it is more intuitive (and robust) to check if NULL keyword is
present in their specs and not analyse the body.

* gnat2why/flow/flow-analysis.adb
(Is_Param_Of_Null_Subp_Of_Generics): check if NULL attached to spec of a
 generic actual subprogram.

Change-Id: Ifd08991e171a2cfa0b532918467b54cc0262849b