Relatively minor change to flow test scripts (use do_flow).
The __flow tests are intended to be fast and mainly about flow. We *do*
actually do VC generation but we really only care if we crash in proof
or not. If a test cares about anything else it should not be __flow.

This update just makes sure we use do_flow everywhere for __flow tests.

Change-Id: I63bdf7d116cb568ae4813649bc63349c93288ef4