P909-030 duplicate checks on split scalar types
Some checks used to be duplicated because, when converting between
two scalar types in split form, we did not change the Ada_Node
associated to the why type. We now introduce a dummy node in this
case so that we have the expected Ada_Node.

* gnat2why/why/why-gen-expr.adb
(Insert_Scalar_Conversion): When converting between types in split
form, introduce a dummy why node to change the associated Ada node.

Change-Id: If2f88d3749da82d8097d5a117665456e949fd04e