P613-025 initial support for extended Ravenscar
Until the support for the GNAT Extended Ravenscar is completed we need a
"GNATprove tasking profile," which is in the middle between extended and
strict Ravenscar. That is what this patch provides.

Some of the GNAT Extended Ravenscar restrictions are special cased, as they
are implied by more conservative restrictions from the old Ravenscar:
- Pure_Barriers is implied by Simple_Barriers,
- No_Implicit_Protected_Object_Allocations and No_Implicit_Task_Allocations
  is implied by No_Implicit_Heap_Allocations.

In effect, if the GNAT configuration file contains:

  pragma Profile (Ravenscar);

then GNATprove will still allow the use of tasking constructs, despite the
fact that some of the GNAT Extended Ravenscar profile restrictions are not
really set.

* gnat2why/spark/spark_definition.adb
(Ravenscar_Profile): replaced with GNATprove_Tasking_Profile with the
 currently required tasking restrictions.

Change-Id: Iee5d47354e1de451276f8a149c3dbf1e5cb9c55b