P714-004 limited support for single file analysis using CP wrapper
Limited support for gnatprove -u switch when CodePeer is used. If only a
single file is passed with -u, CP, when called from SPARK,  only
analyzes that file. When -u is not used, or more than one file is
passed, CP analyzed all files of the project.

* gnatprove.adb
(Run_CodePeer): use option -file of codepeer wrapper
* spark_codepeer_wrapper.adb
(Parse_Command_Line): add option -file
(Generate_SCIL): if -file was given, pass it on to gprbuild

Change-Id: Iae2e1d1ce36212693ad31233ac94c27051456868