P915-004 computation of length on modular
For arrays ranging over modular types, the computation
of the Length attribute is now done on bitvectors when
in the term domain. In the program domain, the
computation remains on mathematical integers to allow
for easy range checks.

* gnat2why/why/gnat2why-expr.adb
(Transform_Attr): Take the expected type as an additional
parameter. Directly translate the compuation of length
attributes on arrays as an operation in the expected type
if we are not in the program domain.

* gnat2why/why/why-gen-arrays.adb
(Get_Array_Attr): Take the why representation type that
should be used for the computation of Length as an
additional argument.
(Build_Length_Expr): Take the why representation type that
should be used for the computation of Length as an
additional argument.

Change-Id: I7fd4bf9a25f27ad84f000abd6f40be2b118582a8