O720-004 do not corrupt ALI with new lines
This test was meant to detect truncated ALI files. Originally this was
simulated with test.cmd driver:

  head -n -4 gnatprove/test.ali > gnatprove/tmp.ali
  rm gnatprove/test.ali
  mv gnatprove/tmp.ali gnatprove/test.ali

but after re-write to test.py we now also intersperse ALI lines with empty
lines. Most likely this is not needed as such a corruption will never happen
in real when gnatprove is interrupted.

* testsuite/gnatprove/tests/O720-004__corrupted_ali_file/test.py
Do not corrupt ALI file with extra empty lines.

Change-Id: I1402898401dde715221ccecc4dbb748ff29a5965