P715-019 Improved conditions on relevant subprograms
We now have more precise conditions to determine which subprograms are
relevant to termination. We also do not register as nonreturning and
relevant to termination predefined subprograms whose scope is an instance
of a generic unit.

* gnat2why/flow/flow-analysis.adb
(Check_Termination): removed local variable

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): improved check on subprograms to register as
nonreturning and relevant to termination.
(Flow_Analyse_Entity): improved the way we check for predefined subprograms

Change-Id: Iafaa0b972fefb355553b0b11c4e7b5d4e3de7eed