cleanup detection of calls to protected operations
Calls to protected entries were registered as both "calling an entry" and
"write-locking of protected object". This was needless and was only making
the ALI file bigger. In phase 2 we collect both calls to entries and to
protected procedures anyway (and protected functions too) in
Directly_Called_Protected_Objects, which is used for ceiling priority
protocol checks.

Also, we were detecting entries based on Convention_Entry, which is
equivalent but less intuitive than Ekind of Entry_Kind.

In preparation for multiple entries and entry families of the GNAT Extended
Ravenscar profile.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Call_Statement): cleanup detection of calls to protected procedures and
 entries.

* gnat2why/flow/flow-analysis.adb
Style fixes in related code.

(no-tn-check)

Change-Id: I3d74e160bd228e43da262198e6a5338222583fa8