O908-020 flow - spurious errors on internal renaming
We now do not generate spurious errors on renamings which are only
internal and not user written. This fixes also O804-004.

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Expression): do the check for N_Object_Renaming_Declaration only if
this comes from source

* testsuite/gnatprove/tests/O804-004__inlined_renamed_actuals_problem
Enable test and update output

* testsuite/gnatprove/tests/spark_book
Enable some files in Chapter-02 and Chapter-08 and update output

Change-Id: I943b374e77534d84264d8448532af56224e40d15