P930-007 protected function on record/array prefix
Read-locks of protected objects are now detected in the same way as
write-locks and do not crash if the protected object is a component of a
record or array.

* gnat2why/flow/flow_utility.adb
(Collect_Functions_And_Read_Locked_POs): detect read-locks of protected
 components just like wrote-locks; local variable renamed from Scope to Scop
 to avoid shadowing of a well-known front end routine.

* testsuite/gnatprove/tests/P930-007__flow_composite_prot_func/
Test output added. Test no longer crashes.

* testsuite/gnatprove/tests/P930-007__flow_composite_prot_func/
Spurious check on internal call to protected function. P930-022.

Change-Id: I9dd50d7b80473b9ac217982073729a1161cab363