PA04-015 spurious DIC check on derived private type
When a derived type of a type with a DIC whose full view is not
in SPARK is of a record kind, a DIC check is emitted for it.
As the fullview of the type is not in SPARK, the DIC check is
spurious.

* gnat2why/spark/spark_util-types.adb
(Needs_DIC_Check): Go through a type hierarchy to decide if an
inherited DIC needs to be checked in SPARK, that is, if the type
on which the DIC was defined is in SPARK or not.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Check): Only issue a DIC check if it is needed.

Change-Id: I1e57ee79e74dbafef4096e3fa00a77ace12c54f7