P715-007 introduce a Terminating GNATprove annotation
That can be used to specify that a subprogram will
always terminate. The information is not used currently
but I have already annotated some tests with it.

* gnat2why/spark/spark_util-subprograms.adb
(Is_Predefined): Check if a subprogram is declared in a
predefined unit or an instance of a predefined unit.

* gnat2why/why/gnat2why-annotate.adb
(Check_Terminate_Annotation): Analyze pragma Annotate
(Gnatprove, Terminating).
(Has_Terminate_Annotation): To be used in flow analysis,
return True if a subprogram has a terminating annotation.
(Syntax_Check_Pragma_Annotate_Gnatprove): Support
terminating annotations.

Change-Id: I3cf368f06a1bdc6ecba7586f2997495113c05fe1