P909-031 temporarily disable test on Darwin
Messages for ceiling priority protocol depend on which ALI files is read
first and this differs between Darwin and Linux. Workaround is easy but will
make implementing the remaining features of the GNAT Extended Ravenscar
profile more difficult.

Disable the test on Darwin just for releasing new version.

* testsuite/gnatprove/tests/P115-039__spark16demo/test.opt
Disable test on Darwin.

Change-Id: I64f715a0ac29fe3588d060f957545a0f43bfcd29