adjust ordering of tasking checks in test.outs
Front end change of how the System package is loaded when it is implicitly
needed (revision 324741, for PA06-029 and P930-047) affected the ordering of
the tasking exclusivity checks. This patch merely fixes one test output.

Note: the ordering of messages should not depend on the order in which
packages are loaded, there is already a TN for this, P909-031.

* testsuite/gnatprove/tests/O917-095__flow_tasking_main_program/test.out
Test output updated.

Change-Id: I74860252b00d1f2d40e5c179754004069d0c4a8d
(cherry picked from commit e42a9886fdb8a7a8b34adea3089ce5e302dcf0f8)