P921-018 use timeout instead of steps in level switch
The level switch aims at giving an easy, high-level way to tune the
verification power of the tool. Levels used to set step limits. This made
the switch rather unpredictable to the user, as steps are not
straightforwardly related to time spent.

The level switch now uses timelimits, 1s for levels 0 and 1, 5s for levels 2
and 3, and 10s for level 4.

As a side effect, the default no longer sets a level, but rather directly
steps so that it remains reproducible.

The 'auto' value for timelimit has also been removed as it has no meaning
anymore.

* gnatprove/configuration.adb
(Set_Level_Timeout_Steps_Provers_Proof_Mode): Set the appropriate timeout
 for the level switch. If no level switch is provided, set the timeout to 0
 and the step limit to 100 (the default). Specific values for steps and
 timeouts override this setting.

Change-Id: Ib7fa6cb7b548d9ea81dcae445673a5c7b3311f8a