P921-029 routine to detect external calls
This patch is follow-up on a cleanup to Get_Enclosing_Concurrent_Objects,
whose result kind is used to find out whether a call to protected subprogram
is external or internal. This functionality should be out-sourced to a
separate routine (because Get_Enclosing_Concurrent_Objects does too much and
is too difficult to use).

It turns out that what we need was already nicely coded in proof and now it
is refactored into a standalone utility routine.

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Is_External_Call): new routine with old code taken from proof; a note added
 about harmlessly incorrect semantics.

* gnat2why/why/gnat2why-expr.adb
(Compute_Param): now uses a refactored routine.

Change-Id: I25fa43968f29badff97b0f7d48e3385018d6da50