PA20-065 add test for gnatprove --help
Trivial test with just a call to "gnatprove --help" to make sure that this
indeed works, to help track of how the help message evolves and to help
detect whitespace issues.

Also, the test.out message is used directly in the UG, which ensures that
the UG contents is always up-to-date.

* testsuite/gnatprove/tests/PA20-065__gnatprove_help/
Added.

* docs/ug/appendix/command_line_invocation.rst
(Command Line Invocation): now automatically inserts the current --help
 output.

Change-Id: I111c0b4e49f18c7ad55bc283fb2b5efd27388d63

Former-commit-id: a50591a9836c49ee8704388cf6183cd36313b93c