P715-019 flow - subprogram termination
We now consider the presence of the terminating annotation. If the
subprogram is not in spark and it is annotated with the terminating
annotation then we assume it is terminating. If the subprogram is in spark
we analyse the code. If there is the terminating annotation and the
subprogram turns out to be potentially nonterminating then we emit a
warning. Now we also do not make any difference between relevant and
non-relevant calls.

* gnat2why/flow/flow-analysis.ads
* gnat2why/flow/flow-analysis.adb
(Check_Terminating_Annotation): checks the consistence between the
terminating annotation and the results from flow analysis.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): we assume subprograms with no body to be
terminating. We don't register anymore subprograms relevant for proving
termination.
(Flow_Analyse_CUnit): check consistency of terminating annotation

* gnat2why/flow/flow_generated_globals-ali_serialization.ads
* gnat2why/flow/flow_generated_globals-ali_serialization.adb
* gnat2why/flow/flow_generated_globals-phase_1.ads
* gnat2why/flow/flow_generated_globals-phase_1.adb
Remove data for registering subprograms relevant to termination

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Add_Edges): we insert in the call graph any callee of the caller but we
do nothing if the caller itself is nonreturning.

* testsuite/gnatprove/tests
Update outputs

Change-Id: Ie1a45b3422b37d72942a1ed34c9033dc1bf6380a