PA18-023 Removed flow-trivia and re-use Generated_Calls
Flow-trivia was only used to provide the set of Direct_Calls to be used in
Gnat2why-Driver. This was not needed and now we instead use
Flow_Generated_Globals-Phase_2.Generated_Calls.

* gnat2why/flow/flow-trivia.ads
* gnat2why/flow/flow-trivia.adb
Removed

* gnat2why/flow/flow_generated_globals-phase_2.ads
* gnat2why/flow/flow_generated_globals-phase_2.adb
(Generated_Calls): defined function to return the list of direct calls

* gnat2why/why/gnat2why-driver.adb
(Generate_Assumptions): replaced Direct_Calls with Generated_Calls

Change-Id: Ib1bb2ccbe2df7f1a028a74f550a316958b1c09a6