P218-017 support --why3-conf in --cache mode
Tests that specify --why3-conf switch did not gain any speed up when running
the testsuite with --cache switch.

It was because the argument to --why3-conf is a full path to the test
sandbox and includes the PID of the test driver process, e.g.
/tmp/tmp-test-P920-047__cvc4_step_consumption-17123/src/why3.conf, so it is
different for every execution of the run-tests script.

Now we recognize the --why3-conf switch and do not hash the full path, but
instead hash the contents of the why3 config file.

* scripts/cache_wrapper
(hash_cmd_line): support --why3-conf switch.

Change-Id: I51a791fd5a3fc6a2df942443784fb4f9018243d4

Former-commit-id: 931c4832d200ec9864171948d0181cd90a4345d1