PA23-006 use correct view to check if a variable is a read effect
We used to check if a variable is read by a subprogram by testing the
presence of a Flow_Id with Variant => Normal_Use in a set that contains only
Flow_Ids with Variant => In_View. In effect, proof did not recognize globals
as being initialized.

* gnat2why/why/gnat2why-util.adb
(Is_Initialized): use In_View, not Normal_View in container membership test.

* testsuite/gnatprove/tests/K518-029__renaming/test.out
Counterexample changed.

* testsuite/gnatprove/tests/N513-030__dyn_bounds_in_loop/test.out
New checks proved (global that is an IN parameter of the enclosing
subprogram is now correctly recognized as initialized).

* testsuite/gnatprove/tests/NC06-004__init/test.py
Increase steps; needed to keep checks provable.

Change-Id: I6d5118b16ce1137f5ce9ab5bf893ec15f66ff767

Former-commit-id: 2b65cd4e1a51c78beb9140e866f2411d83c3fb27