PA14-032 assume bounds in axiom of array aggregates
Array aggregates are translated as an uninterpreted function with
an axiom stating the values of its content. If the aggregate is
of a dynamic array type and it is not a 'Update attribute, the
bounds of the aggregate were not assumed, which resulted in
unprovable checks whenever the check applied directly on an
aggregate. We now assume the bounds of array aggregates.

* gnat2why/why/gnat2why-expr.adb
(Transform_Aggregate): Assume the value of the aggregate bounds.
As they may depend on variables, the bounds are given as additional
parameters to the aggregate function. A guard is added to the axiom
stating the dynamic property of the bounds to avoid assuming an
unsound axiom if they bound fail a runtime check.

* gnat2why/why/why-gen-arrays.adb
(New_Dynamic_Property): Function that creates a call to an array
dynamic property.

* gnat2why/why/why-gen-arrays.ads
(New_Dynamic_Property): Calls Gen.Arrays.New_Dynamic_Property on
arrays.

Change-Id: Ib04616545f6b28bc5252776a70b32cfb89ad2e83

Former-commit-id: 3e43044280a10c820db2736ebf683c67dfcf6b3c