OA23-036 counterex - Printing arrays of records
Adding counterexamples for arrays of records. These were not printed as
the projection mechanism is quite different. New arrays are used as
projections for each field.
In this commit, we created functions that take values of arrays
corresponding to each field to rebuild a correct value of array of
records.
Also updated outputs for recent commit on P419-017, PA03-023
and OA23-036 all on counterexamples.

* testsuite/gnatprove/tests/*/test.out
Updated output.
* testsuite/gnatprove/tests/text_io_get_line/test.opt
Removed XFAIL as part of PA03-023.

* gnat2why/why/gnat2why-counter_examples.adb
(Array_Record__Value): Added a container for array of records values.
(Add_V): Add a new value to an array of records without checking for its
existence.
(Trimi): Unfound function trim that is supposed to remove each occurence
of a character in a string.
(Pretty_Print_Element): Pretty print pair of element for future display.
(Non_Empty_Split): Copy of String_Split.Create which remove the specific
case when no separators is encountered. From testing, it was getting
no slice and we wanted to get a singleton with the initial string.
(Merge): Correctly add the value of an array for a given field to the rest
of the array of records.
(Get_List): Create an array of record of value for a given field then
merge it to its first argument.
(Print_Array_Into_Value): Correctly print an array of record from the
structure Array_Record_Value.Map.
(Get_CNT_Element_Values_And_Attributes): Adding a case for arrays which
contains fields (in the why3 output): arrays of records. Iterate over
the fields given by why3 and not on the Node's field because the node is
an array (bad for the order of the fields).

Change-Id: I671ab6792fac0aae584ff0a2530ac643fcb3199f

Former-commit-id: 43623e87a16e26bc4c867214182d9ee78e6bb0da