P715-019 nonreturning predefined subprograms
Routine Is_Predefined must be called on the subprogram entity, as it is said
in its comment, not on the subprogram's scope entity. Otherwise it gives
wrong result.

This requirements is now enforced by the contract.

* gnat2why/spark/spark_util-subprograms.ads
(Is_Predefined): contract added.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): adjusted to match the contract of Is_Predefined.

Change-Id: I7ceb9a37ca08224aa97554715ad094fa7abd5700

Former-commit-id: 81b9fbb0d5f7a79aaaa6a6bfa17f7d449b0daa2f