PA27-004 Fix broken toctree for SPARK UG
The use of conditional compilation of the SPARK UG based on a core/full
mode (passed as tags to Sphinx commands) does not work for conditionally
including or not sections in the toctree. Instead, comment out the toctree
for the core mode, which needs to be commented out for generating the
reduced version of the SPARK UG suitable for translation in other
languages.

Change-Id: I6527293934b297840bde20d1b08a04582000b18c

Former-commit-id: 5f0bf79ffccbd12678c0da8be50973ce865cdbc0