P828-003 fix annotate aspect
The case where an Annotate aspect was used instead of pragma Annotate
was broken. Internally, such an aspect generates a pragma, but with an
extra argument, which is the entity. This makes the strict syntax
checking fail.

We fix it by taking into account explicitly this case in the syntax
check.

* gnat2why-annotate.adb
(Syntax_Check_Pragma_Annotate_Gnatprove): take into account the case
where the pragma has 5 arguments, and the last one is an entity
argument.

Change-Id: I1d2e59ab77a6887c7213833198a538156070ef9f

Former-commit-id: 625081dc7ec7cb65bb12e24ca116fd42b1c2c062