PA26-037 Allow nested loop before loop (in)variant
Originally, nested loops were rejected by GNATprove to avoid the
declaration of non-scalar objects before the loop in(variant), as the
translation to Why3 cannot handle these in case they lead to the
definition of constants (either for the object or its bounds or
discriminants).

It was also a way to avoid the possible inefficient duplication of
generated code between the initial and the looping part of the outter
loop translation, as both will contain the nested loop.

This limitation is now lifted, to allow nested loops before the loop
(in)variant, by traversing the loop statements to detect non-scalar
object declarations. The code duplication is no worse than the code
duplication for non-looping code before the loop (in)variant. In theory
one could build a sequence of nested loops that lead to quadratic
explosion, but this is highly improbable, and the workaround would be
simply to move the loop (in)variant(s) before the nested loop(s).

* gnat2why/spark/spark_definition.adb
(Check_Loop_Invariant_Placement): Do not restrict nested loops anymore.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Allow the presence of nested loops before
  the (in)variant.

* docs/ug/appendix/gnatprove_limitations.rst
Update limitation.

* tests/M301-012__array_of_iter/test.out
* testsuite/gnatprove/tests/N731-009__not_supported_loops/test.out
No more messages about nested loop restriction.

* testsuite/gnatprove/tests/PA26-037__loopinv/lt.adb
* testsuite/gnatprove/tests/PA26-037__loopinv/lt.ads
* testsuite/gnatprove/tests/PA26-037__loopinv/test.out
* testsuite/gnatprove/tests/PA26-037__loopinv/test.py
* testsuite/gnatprove/tests/PA26-037__loopinv_variants/simple_loops.adb
* testsuite/gnatprove/tests/PA26-037__loopinv_variants/test.out
* testsuite/gnatprove/tests/PA26-037__loopinv_variants/test.py
New tests.

Change-Id: Ibe09f48a53b58a60b3df9383325714f11d515971

Former-commit-id: 3e9a5ad2070af79bb3ead8b9f5e45632a8daec53