P921-029 implicit PO are not formals
One of the final cleanups related to Get_Enclosing_Concurrent_Object.

Protected objects that appear as prefixes of external calls to protected
subprograms are now handled as actuals, not as formals. Protected types that
appear as prefixes to internal calls are now ignored: these calls behave
like calls to nested subprograms (i.e. no new "global" objects gets visible
to the callee and no locks are acquired, btw.).

Previously internal calls introduced an implicit formal of the enclosing
protected type or of the anonymous protected object (for single protected
objects). This was quite odd and was making the code really confusing,
because the "enclosing concurrent object" could be either a type or an
object.

* gnat2why/flow/flow-interprocedural.adb
(Add_Simple_Procedure_Dependency): Include, not Insert, the protected
 object and explain why.

* gnat2why/flow/flow_utility.ads
(Get_Implicit_Formal): Callsite and Entire parameters removed.
(Get_Formals): Callsite and Entire parameters removed.

* gnat2why/flow/flow_utility.adb
(Get_Implicit_Formal): now always returns a protected type and never a
 (single) protected object.

(Get_Formals): Callsite and Entire parameters are no longer propagated to
 Get_Implicit_Formal.

* testsuite/gnatprove/tests/OA23-011__flow_garbage_tk_suffix/
Disabled. It wrongly rejected by the flow analysis (that complained about
a missing Global where Globals have been in fact generated by the tool).
Now the code is adjusted so that there is indeed missing global, but we are
failing to detect this, because front end attaches Globals contracts to the
single task object and not to the single task type (where we expect it).
Should be again rejected when O527-025 is done.

Change-Id: Ia2c0834cf0684d5824637fd4d08a657a2516cead

Former-commit-id: 56af0717462ae0fa40e95d75b1982c9cccb66d30