PB02-024 Part_Of variables are not automatically volatile
Variables that are Part_Of single concurrent objects are indeed volatile
from the outside of the concurrent object, but it does not really matter,
because they cannot be accessed from the outside.

From the inside of concurrent object they are non-volatile (unless
explicitly annotated as Volatile).

* gnat2why/spark/spark_util.adb
(Has_Volatile): Part_Of variables are not automatically volatile.

* testsuite/gnatprove/tests/O429-034__flow_concurrent_parts/test.out
Check on a Part_Of accessed within task body is now provable; update output.

Change-Id: Id587913ab66233e4ce004b6aee345de61cbce00a

Former-commit-id: a16e3e9cc8521717bb0be21e7ecd25fd3bac7bd8