P921-029 rename Get_Enclosing_Concurrent_Object
Routine Get_Enclosing_Concurrent_Object returned either a simple concurrent
object (if possible) or a concurrent type (otherwise), so the name was
terribly confusing.

The only plausible rationale was that when we analyse single concurrent
types we want to represent the type by a unique Entity_Id. Using the
Entity_Id of the single concurrent object guaranteed that globals arriving
via external calls use the same Entity_Id (i.e. protected subprogram of a
single protected object calls a proxy that calls another protected
subprogram of the same single protected object). But such an external call
is not allowed in SPARK, because it is potentially blocking and we check for
this.

So we were using the Entity_Id of the single concurrent *object* if possible
for no real reason and at the same time we were calling
Get_Enclosing_Concurrent_Object to get the Entity_Id of a concurrent *type*
(even if no objects of this type exist at all).

Now we always represent enclosing concurrent types with the type entity and
this is reflected both in the routine name and its contract.

* gnat2why/flow/flow-types.ads
(Enclosing_Concurrent_Type): rename from Get_Enclosing_Concurrent_Object.

* gnat2why/flow/flow-types.adb
(Enclosing_Concurrent_Type): always return type entity.

* gnat2why/flow/flow_utility.adb
(Up_Project_Constituent): remove assertion which now merely duplicates the
 contract of Enclosing_Concurrent_Type.

Change-Id: I57a58b7d123e6ac4de1ca337c8b118d8279f8e58

Former-commit-id: 6c70be0c427688d777c3d61cd18412a450eef200