P715-019 Save info on Terminating annotation in ALI
We now save subprograms annotated with the Terminating annotation to the
ALI file. Because of this we can re-use this information in phase 2 when
deciding which subprograms are potentially nonreturning. In particular if
a subprogram is annotated with Terminating and we are analysing it, we
still issue a warning in case we find it to be potentially nonreturning,
but if we are seeing it as a callee of another subprogram then we trust
its annotation.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): register subprograms with Terminating annotation

* gnat2why/flow/flow_generated_globals-ali_serialization.ads
* gnat2why/flow/flow_generated_globals-ali_serialization.adb
* gnat2why/flow/flow_generated_globals-phase_1.ads
* gnat2why/flow/flow_generated_globals-phase_1.adb
Add structure to store subprograms with Terminating annotation

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Add_Edges): do not seed the graph with subprograms annotated with
Terminating

(Calls_Potentially_Nonreturning_Subprogram): only check callees of a
subprogram if they are not annotated with Terminating annotation

(Merge): added to avoid repeated code.

Change-Id: I9e4c11b8371ee3bcc7c876da98056f39f7712c39

Former-commit-id: efe155287da7e84e845e3a5e6a33b87e50d0bd9f